Please Invest on your Risk
This code continuously retrieves the current market data for the list of stocks defined in the stocks variable and calculates the daily percentage change for each stock. It then identifies potential trending stocks based on the mean and standard deviation of the percentage change and the current percentage change in each stock. The list of trending stocks is printed to the console every minute.

Note: This code is for educational purposes only and should not be used for actual investment decisions.